+++ CHANGELOG.md
@@ -31,7 +31,7 @@
 - Fixed Ticket 'Submission Update' workflow to avoid possible looping.
 - Moved misleading filter icons from CustomTable headers to last table header.
 - Adjusted positioning on CustomTable Settings to prevent overflowing offscreen 
-- Fixed overflowing table styles from Reports/WeeklyReportTemplate.jsx
+- Fixed overflowing table styles from Reports/ReportTemplate.jsx
 
 ## [0.3.1] - 2022-06-06